+++ src/main/java/com/smashingmods/alchemylib/api/blockentity/container/FakeItemRenderer.java
@@ -1,30 +1,25 @@
 package com.smashingmods.alchemylib.api.blockentity.container;
 
-import com.mojang.blaze3d.systems.RenderSystem;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.world.item.ItemStack;
 
 @SuppressWarnings("unused")
 public class FakeItemRenderer {
 
-    public static void renderFakeItem(GuiGraphics pGuiGraphics, ItemStack pItemStack, int pX, int pY, float pOpacity) {
-        renderFakeItem(pGuiGraphics, pItemStack, pX, pY, pOpacity, false);
+    public static void renderFakeItem(GuiGraphics pGuiGraphics, ItemStack pItemStack, int pX, int pY) {
+        renderFakeItem(pGuiGraphics, pItemStack, pX, pY, true, false);
     }
 
     public static void renderFakeItem(GuiGraphics pGuiGraphics, ItemStack pItemStack, int pX, int pY, boolean pDrawItemDecorations) {
-        renderFakeItem(pGuiGraphics, pItemStack, pX, pY, 1f, pDrawItemDecorations);
+        renderFakeItem(pGuiGraphics, pItemStack, pX, pY, true, pDrawItemDecorations);
     }
 
-    public static void renderFakeItem(GuiGraphics pGuiGraphics, ItemStack pItemStack, int pX, int pY, float pOpacity, boolean pDrawItemDecorations) {
-        if (pOpacity != 1f) {
-            RenderSystem.setShaderColor(1f, 1f, 1f, pOpacity);
-        }
-
+    public static void renderFakeItem(GuiGraphics pGuiGraphics, ItemStack pItemStack, int pX, int pY, boolean pSemiTransparent, boolean pDrawItemDecorations) {
         pGuiGraphics.renderFakeItem(pItemStack, pX, pY);
-
-        if (pOpacity != 1f) {
-            RenderSystem.setShaderColor(1f, 1f, 1f, 1f);
+        if (pSemiTransparent) {
+            pGuiGraphics.fill(RenderType.guiGhostRecipeOverlay(), pX, pY, pX + 16, pY + 16, 0x88888888);
         }
 
         if (pDrawItemDecorations) {